+++ src/init.c
@@ -124,7 +124,6 @@ static const R_CallMethodDef callMethods[] = {
     {"Cinrange", (DL_FUNC)&inrange, -1},
     {"Cbetween", (DL_FUNC)&between, -1},
     {"ChasOpenMP", (DL_FUNC)&hasOpenMP, -1},
-    {"CbeforeR340", (DL_FUNC)&beforeR340, -1},
     {"CuniqueNlogical", (DL_FUNC)&uniqueNlogical, -1},
     {"CfrollfunR", (DL_FUNC)&frollfunR, -1},
     {"CdllVersion", (DL_FUNC)&dllVersion, -1},
@@ -396,18 +395,6 @@ SEXP hasOpenMP(void)
   return ScalarInteger(0);
 #endif
 }
-
-SEXP beforeR340(void)
-{
-// used in onAttach.R for message about fread memory leak fix needing R 3.4.0
-// at C level to catch if user upgrades R but does not reinstall data.table
-#if R_VERSION < R_Version(3, 4, 0)
-  return ScalarLogical(true);
-#else
-  return ScalarLogical(false);
-#endif
-}
-// # nocov end
 
 extern int *_Last_updated; // assign.c